+++ TestProjects/VisualEffectGraph/Assets/AllTests/Editor/Tests/VFXPrefabTests.cs
@@ -147,7 +147,8 @@ namespace UnityEditor.VFX.Test
                 parameter.value = i + 1;
                 graph.AddChild(parameter);
             }
-            graph.RecompileIfNeeded();
+
+            AssetDatabase.ImportAsset(AssetDatabase.GetAssetPath(graph));
 
             var mainObject = MakeTemporaryGameObject();
             var vfx = mainObject.AddComponent<VisualEffect>();
@@ -267,7 +268,7 @@ namespace UnityEditor.VFX.Test
             {
                 Add_Valid_System(graph);
             }
-            graph.RecompileIfNeeded();
+            AssetDatabase.ImportAsset(AssetDatabase.GetAssetPath(graph));
 
             var mainObject = MakeTemporaryGameObject();
             GameObject prefabInstanceObject;
@@ -346,7 +347,8 @@ namespace UnityEditor.VFX.Test
             parameter.SetSettingValue("m_Exposed", true);
             parameter.value = new Vector3(0, 0, 0);
             graph.AddChild(parameter);
-            graph.RecompileIfNeeded();
+
+            AssetDatabase.ImportAsset(AssetDatabase.GetAssetPath(graph));
 
             var mainObject = MakeTemporaryGameObject();